+++ pycondor/__init__.py
@@ -2,7 +2,8 @@
 
 from __future__ import division, print_function
 
-from .pycondor import Job, Dagman
+from .job import Job
+from .dagman import Dagman
 from .base import get_queue
 
 __version__ = '0.1.1'